workspace include watcher scripts
Some convenient scripts for when you just want to be able
to trivially see changes as they happen.

ie, watch-git_status is typically running in one of the jumble
of terminals I have for my currently active project. Yes, its
info is available in vscode, but it is easier to view like this.

These live in gopath beside goto.sh